requirements.txt
django>=4.2
requests>=2.28
python-decouple>=3.6
celery>=5.2
django-celery-beat>=2.5
redis>=4.5
django-redis>=0.6
nltk>=3.8
textblob>=0.17.1
gunicorn>=21.0
whitenoise>=6.0
dj-database-url>=2.0